refactor: move TanStack Query hooks from demos.ts to dedicated hooks.ts file
